modularized display code for agents; updated unit tests accordinly
