+++ SRAgent/agents/entrez.py
@@ -17,7 +17,7 @@ from SRAgent.agents.esearch import create_esearch_agent
 from SRAgent.agents.esummary import create_esummary_agent
 from SRAgent.agents.efetch import create_efetch_agent
 from SRAgent.agents.elink import create_elink_agent
-from SRAgent.agents.utils import create_step_summary_chain
+from SRAgent.agents.display import create_step_summary_chain
 
 # functions
 def create_entrez_agent(
