+++ SRAgent/settings.yml
@@ -3,7 +3,7 @@ default:
   db_port: 5432
   db_user: "postgres"
   db_timeout: 30
-  service_tier: "flex"
+  service_tier: "standard"
 
 test:
   db_name: "sragent-test"
@@ -26,6 +26,7 @@ test:
     entrez_convert: "gpt-4.1-mini"
     find_datasets: "gpt-4.1-mini"
     step_summary: "gpt-4.1-mini"
+    get_entrez_ids: "gpt-4.1-mini"
   temperature:
     default: 0.1
     esearch: 0.1
@@ -45,6 +46,7 @@ test:
     entrez_convert: 0.1
     find_datasets: 0.1
     step_summary: 0.1
+    get_entrez_ids: 0.0
   reasoning_effort:
     default: "low"
     esearch: "low"
@@ -64,6 +66,7 @@ test:
     entrez_convert: "low"
     find_datasets: "low"
     step_summary: ""
+    get_entrez_ids: "low"
   service_tier:
     default: "flex"
     esearch: "default"
@@ -83,6 +86,7 @@ test:
     entrez_convert: "default"
     find_datasets: "default"
     step_summary: "default"
+    get_entrez_ids: "default"
     
 prod:
   db_name: "sragent-prod"
@@ -105,6 +109,7 @@ prod:
     entrez_convert: "gpt-4.1-mini"
     find_datasets: "gpt-4.1-mini"
     step_summary: "gpt-4.1-mini"
+    get_entrez_ids: "gpt-4.1-mini"
   temperature:
     default: 0.1
     esearch: 0.1
@@ -124,6 +129,7 @@ prod:
     entrez_convert: 0.1
     find_datasets: 0.1
     step_summary: 0.1
+    get_entrez_ids: 0.0
   reasoning_effort:
     default: "low"
     esearch: "low"
@@ -143,6 +149,7 @@ prod:
     entrez_convert: "low"
     find_datasets: "low"
     step_summary: ""
+    get_entrez_ids: "low"
   service_tier:
     default: "flex"
     esearch: "default"
@@ -162,6 +169,7 @@ prod:
     entrez_convert: "default"
     find_datasets: "default"
     step_summary: "default"
+    get_entrez_ids: "default"
   
 claude:
   db_name: "sragent-test"
