+++ SRAgent/workflows/find_datasets.py
@@ -57,7 +57,7 @@ class EntrezInfo(BaseModel):
     database: str
 
 def create_get_entrez_ids_node() -> Callable:
-    model = set_model(model_name="gpt-4.1-mini", temperature=0)
+    model = set_model(agent_name="get_entrez_ids")
     async def invoke_get_entrez_ids_node(state: GraphState, config: RunnableConfig) -> Dict[str, Any]:
         """
         Structured data extraction of Entrez IDs from message
